feature(helper): add algolia_search() and algolia_search_cdn()
